ref(Backend): Refactored controllers for tutor service
