+++ Backend/FishingBooker/FishingBooker/src/main/java/com/isa/FishingBooker/controller/tutorservice/TutorServicesAbstractController.java
@@ -1,0 +1,20 @@
+package com.isa.FishingBooker.controller.tutorservice;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+
+import com.isa.FishingBooker.model.User;
+import com.isa.FishingBooker.security.auth.TokenBasedAuthentication;
+import com.isa.FishingBooker.service.TutorServicesService;
+
+public class TutorServicesAbstractController {
+	@Autowired
+	protected TutorServicesService tutorServicesService;
+	protected User getLoggedinUser() {
+		TokenBasedAuthentication auth = (TokenBasedAuthentication) SecurityContextHolder.getContext()
+				.getAuthentication();
+		User tutor = (User) auth.getPrincipal();
+		return tutor;
+	}
+
+}
